Extract `dispute_early_close` tests
