+++ zrml/prediction-markets/src/tests/mod.rs
@@ -24,14 +24,15 @@ mod approve_market;
 mod buy_complete_set;
 mod create_market;
 mod dispute;
+mod dispute_early_close;
 mod edit_market;
 mod manually_close_market;
 mod on_initialize;
 mod on_market_close;
 mod reject_market;
 mod report;
-mod sell_complete_set;
 mod schedule_early_close;
+mod sell_complete_set;
 
 use crate::{mock::*, AccountIdOf, AssetOf, BalanceOf, Config, Error, Event};
 use core::ops::Range;
